+++ ui/ceval/package.json
@@ -17,13 +17,26 @@
   "dependencies": {
     "@badrap/result": "^0.2.13",
     "chessops": "^0.12.7",
+    "common": "workspace:*",
     "idb-keyval": "^6.2.0",
     "snabbdom": "^3.5.1",
-    "common": "workspace:*"
+    "stockfish-mv.wasm": "^0.6.1",
+    "stockfish-nnue.wasm": "1.0.0-1946a675.smolnet",
+    "stockfish-web": "^0.0.7",
+    "stockfish.js": "^10.0.2",
+    "stockfish.wasm": "^0.10.0"
   },
   "scripts": {
     "compile": "tsc",
     "dev": "tsc",
     "prod": "tsc"
+  },
+  "lichess": {
+    "copy": [
+      {
+        "src": "node_modules/stockfish*/*.{js,wasm}",
+        "dest": "../../public/npm"
+      }
+    ]
   }
 }
