+++ ui/ceval/src/util.ts
@@ -1,5 +1,5 @@
 import { CevalTechnology } from './platform';
-import { ExternalEngine } from './worker';
+import { ExternalEngine } from './engines/externalEngine';
 
 export function isEvalBetter(a: Tree.ClientEval, b: Tree.ClientEval): boolean {
   return a.depth > b.depth || (a.depth === b.depth && a.nodes > b.nodes);
@@ -24,15 +24,15 @@ export const pow2floor = (n: number) => {
   return pow2;
 };
 
-export const sharedWasmMemory = (initial: number, maximum: number): WebAssembly.Memory => {
+export const sharedWasmMemory = (lo: number, hi = 32767): WebAssembly.Memory => {
+  let shrink = 4;
   while (true) {
     try {
-      return new WebAssembly.Memory({ shared: true, initial, maximum });
+      return new WebAssembly.Memory({ shared: true, initial: lo, maximum: hi });
     } catch (e) {
-      if (e instanceof RangeError) {
-        if (initial === maximum) throw e;
-        maximum = Math.max(initial, Math.floor(maximum / 2));
-      } else throw e;
+      if (hi <= lo || !(e instanceof RangeError)) throw e;
+      hi = Math.ceil(hi - hi / shrink);
+      shrink = shrink === 4 ? 3 : 4;
     }
   }
 };
@@ -51,15 +51,21 @@ export function defaultDepth(technology: CevalTechnology, threads: number, multi
   }
 }
 
-export function engineName(technology: CevalTechnology, externalEngine?: ExternalEngine): string {
+export function engineName(
+  technology: CevalTechnology,
+  officialStockfish: boolean,
+  externalEngine?: ExternalEngine,
+): string {
   if (externalEngine) return externalEngine.name;
   switch (technology) {
     case 'wasm':
     case 'asmjs':
       return 'Stockfish 10+';
     case 'hce':
-      return 'Stockfish 11+';
+      return officialStockfish ? 'Stockfish 16' : 'Stockfish MV';
+    case 'nnue':
+      return 'Stockfish 16';
     default:
-      return 'Stockfish 14+';
+      return 'Error';
   }
 }
