+++ ui/site/package.json
@@ -23,10 +23,6 @@
     "hopscotch": "^0.3.1",
     "lichess-pgn-viewer": "^2.0.0",
     "prop-types": "^15.8.1",
-    "stockfish-mv.wasm": "^0.6.1",
-    "stockfish-nnue.wasm": "1.0.0-1946a675.smolnet",
-    "stockfish.js": "^10.0.2",
-    "stockfish.wasm": "^0.10.0",
     "tablesort": "^5.3.0",
     "voice": "workspace:*",
     "zxcvbn": "^4.4.2"
@@ -86,10 +82,6 @@
       {
         "src": "node_modules/@yaireo/tagify/dist/tagify.min.js",
         "dest": "../../public/npm/tagify"
-      },
-      {
-        "src": "node_modules/stockfish*/*.{js,wasm}",
-        "dest": "../../public/npm"
       },
       {
         "src": "../../node_modules/chessground/dist/chessground.min.js",
